only destroy trade item

+++ app/jobs/trade/trade_item_clean_job.rb
@@ -2,7 +2,7 @@ module Trade
   class TradeItemCleanJob < ApplicationJob
 
     def perform(trade_item)
-      trade_item.destroy
+      trade_item.destroy if ['init', 'checked'].include?(trade_item.status)
     end
 
   end
